ready for image creation

+++ meta-openpli/recipes-openpli/enigma2-plugins/enigma2-plugin-extensions-openwebif.bbappend
@@ -1,26 +1,11 @@
-PRINC = "2"
-
 python do_package_prepend () {
 	boxtypes = [
-		('dm500hd', 'dm500hd.jpg', 'dm_normal.png'),
-		('dm7020hd', 'unknown.jpg', 'dm_normal.png'),
-		('dm8000', 'dm8000.jpg', 'dm_normal.png'),
-		('dm800se', 'dm800se.jpg', 'dm_normal.png'),
-		('et5x00', 'et5x00.jpg', 'et_rc5_normal.png'),
-		('et6x00', 'et5x00.jpg', 'et_rc5_normal.png'),
-		('et9x00', 'et9x00.jpg', 'et_rc7_normal.png'),
-		('et4x00', 'unknown.jpg', 'et_rc13_normal.png'),
-		('xp1000', 'unknown.jpg', 'xp_rc14_normal.png'),
-		('vuduo', 'duo.jpg', 'vu_normal.png'),
-		('vusolo', 'solo.jpg', 'vu_normal.png'),
-		('vusolo2', 'solo2.jpg', 'vu_ultimo.png'),
-		('vuultimo', 'ultimo.jpg', 'vu_ultimo.png'),
-		('vuuno', 'uno.jpg', 'vu_normal.png'),
+		('azboxme', 'me.jpg', 'me.png'),
+		('azboxminime', 'minime.jpg', 'me.png'),
+		('azboxhd', 'premium.jpg', 'premium.png'),
 	]
 	import os
 	top = '${D}${PLUGINPATH}/public/images/'
-	target_box = 'unknown.jpg'
-	target_remote = 'ow_remote.png'
 	for x in boxtypes:
 		if x[0] == '${MACHINE}':
 			target_box = x[1]
@@ -29,9 +14,17 @@ python do_package_prepend () {
 	for root, dirs, files in os.walk(top + 'boxes', topdown=False):
 		for name in files:
 			if target_box != name and name != 'unknown.jpg':
-				os.remove(os.path.join(root, name))
+				if target_box == 'premium.jpg':
+					if not (name == 'elite.jpg' or name == 'premium+.jpg' or name == 'ultra.jpg'):
+						os.remove(os.path.join(root, name))
+				else:
+					os.remove(os.path.join(root, name))
 	for root, dirs, files in os.walk(top + 'remotes', topdown=False):
 		for name in files:
 			if target_remote != name and name != 'ow_remote.png':
-				os.remove(os.path.join(root, name))
+				if target_remote == 'premium.png':
+					if not (name == 'elite.png'):
+						os.remove(os.path.join(root, name))
+				else:
+					os.remove(os.path.join(root, name))
 }
